src/requirements.txt
fastapi~=0.115.12
uvicorn~=0.22.0
pydantic~=2.4.0
pydantic[email]
SQLAlchemy~=2.0.20
pandas~=2.2.3
joblib
scikit-learn
matplotlib
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
